requirements.txt
pandas
numpy
swanlab
wandb
pyyaml